requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.1
Django==5.2.1
djangorestframework==3.16.0
flake8==7.3.0
iniconfig==2.1.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
sqlparse==0.5.3
